requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
psycopg2-binary
python-dotenv
cryptography
email-validator
requests
